Split up oggex-bin into ogx.h, main.cpp
The rationale for doing this is that we can now have a common interface
to use in the GUI app as well as the CLI application.
